Start of my project
I have some very loose starter code so far. I have made a basic outline for what I think the layout of my project will be. There has been no unittesting so far since none of the code so far works for sure, but I can go through the text file and find and parse the rules that match whatever the name of the starter_variable input by the user was.