Choosing a random option
I created a function called generate_random_number in my Rule class which has a way of accounting for the weights of the options in a grammar and choosing a random option every time. I tested this manually to make sure it works and it does. I have no unit tests yet because I'm going to get a couple more things set up in my code before I make unit testing for it.