Rule class
I created a generate_sentence_fragment method in my Rule class that works with the generate_random_number method. This class picks out the option within the rule in a random fashion. I created a unit test for this implementation and I tested by hand to make sure that it worked.