Option Class
I implemented my Option class with a generate_sentence_fragment method that uses the duck typing functionality. The option class goes through whatever option was chosen symbol by symbol and checks if the symbol is a variable or terminal. Whichever one it is, the method sends it to the corresponding method in the class. I created a unit test for the option implementation and I am on stable ground.