Updated my terminal return
I updated my terminal return because I realized that I was returning a list with a duplicate of the terminal being returned. So I only needed to return a list with that one thing in it, so now my indexing is a little easier to understand. I also created some more unit tests to test the duck typing function and the terminal class method.